Changed test.cpp

+++ LR4/calc.cpp
@@ -14,7 +14,7 @@ int main()
 		std::cout << num_1 << " - " << num_2 << " = " << num_1 - num_2 << std::endl;
 		std::cout << num_1 << " * " << num_2 << " = " << num_1 * num_2 << std::endl;
 		std::cout << num_1 << " / " << num_2 << " = " << num_1 / num_2 << std::endl;
-		std::cout << num_1 << ".0 / " << num_2 << ".0 = " << double(num_1) / num_2 << std::endl;
+		std::cout << num_1 << ".0 / " << num_2 << ".0 = " << double(num_1) / double(num_2) << std::endl;
 		std::cout << num_1 << " % " << num_2 << " = " << num_1 % num_2 << std::endl;
 		std::cout << num_1 << " + " << num_2 << " * " << num_3 << " = " << num_1 + num_2 * num_3 << std::endl;
 		std::cout << "(" << num_1 << " + " << num_2 << ") * " << num_3 << " = " << (num_1 + num_2) * num_3 << std::endl;
